Avoid using readElementText()
As we read the XML incrementally as we receive it over the network, it
is important that we can resume parsing even if data is split in the
middle of a tag.
Using readElementText() is detrimental for this, as it cannot backtrack
if the network packet split happens to be inside the element text.

This caused some weird issues, where tags with id 0 were inserted,
because the element text was invalid due to the split, and toInt() just
silently returned 0. Since readElementText() cannot backtrack, the
invalid value would also not be overwritten once the second network
packet arrived.

Now we use a reverse QXmlStreamReader::Characters parsing style, which
will only be triggered once the whole text has been read, thus there is
no need to backtrack. The new pain now is that this method regularly
returns meaningless whitespace strings, which we have to guard against
with an extra isWhitespace() check.