React to Android intents
This is utterly cursed, especially because we cannot react to the intent
right away, but have to wait for the Qt thread to start up until the
QGuiApplication is created.

Even though the C++ <-> Java interface was improved recently [0], the
whole FFI is just ugly as fuck [1].

Following the recommendation [2] we use QtActivityBase as the base class
(which is almost the same as QtActivity) and override the needed intent
methods.

To make matters worse, for some reason content URIs do not work with the
PDF view, even though there is explicit support in QML [3].

This endeavour should finally allow us to open PDF files from other
apps.

[0] https://www.qt.io/blog/unstringifying-android-development-with-qt-6.4
[1] https://www.kdab.com/qt-android-episode-5/
[2] https://www.qt.io/blog/changes-to-android-java-bindings-qt-6.7
[3] https://www.volkerkrause.eu/2019/02/16/qt-open-files-on-android.html