Always use Image for displaying PDFs
It seems the QML Pdf views are currently very buggy, especially when it
comes to zooming (it just does not allow to zoom, and always resets back
the renderScale value to the default).
Luckily QtPDF installs a QImageIOHandler, which makes PDFs usable from
Image as well, and it seems less buggy.

In the future we might want to reevaluate this, if we need more
features, but for now it should be fine.